+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - ui: new output for `backtrace` command (with source file and line)
+- debugger: add `--save-history` option
 
 ### Changed